backend-python/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
mutagen==1.47.0
pypdf==4.2.0
Pillow==10.3.0
minio==7.2.7
pyexiftool==0.5.6
